requirements.txt
amadeus==8.1.0
python-dotenv==1.0.0
mistralai==0.0.12
schedule==1.2.1
loguru==0.7.2
PyYAML==6.0.1 